refactor: 모듈화 적용
main.js 모듈화
