+++ src/constants.js
@@ -1,0 +1,21 @@
+export const ANIMATION_DURATION = 700; // 변경 가능
+export const DIGIT_MIN = 0;
+export const DIGIT_MAX = 9;
+export const USER_INPUT = "#user-input";
+export const SUBMIT_BUTTON = "#submit-button";
+export const RESULT_BOARD = "#result-board";
+export const RESULT_HEAD = "#result-head";
+export const BACKSPACE = 8;
+export const ENTER = 13;
+export const STRIKE_POINT = 10;
+export const BALL_POINT = 1;
+export const BOARD_HEADER_CORRECT = "&#127881;정답을 맞추셨습니다&#127881;";
+export const BOARD_PARAGRAPH_CORRECT = "게임을 새로 시작하시겠습니까?";
+export const BOARD_HEADER_WRONG = "&#128517;틀렸습니다&#128517;";
+export const BOARD_HEADER_READY = "숫자를 제출해 주세요";
+export const RESET_BUTTON_CONTENT = "게임 재시작";
+export const ALERT_DIGITS_NOT_LONG_ENOUGH = "최소 3자리의 숫자를 입력해 주세요";
+export const ALERT_DIGITS_NOT_DISTINCT = "중복되는 숫자 없이 주세요";
+export const NOT_FOUND_VALUE = -1
+export const ARROW_LEFT = 37;
+export const ARROW_RIGHT = 39;
